requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
cmudict==1.0.2
discord==1.7.3
discord.py==1.7.3
emoji==1.6.1
html5lib==1.1
idna==3.3
lorem==0.1.1
multidict==5.2.0
Pillow==9.1.0
pronouncing==0.2.0
pycorpora==0.1.2
python-dotenv==0.19.2
requests==2.27.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
tracery==0.1.1
typing_extensions==4.0.1
urllib3==1.26.8
webencodings==0.5.1
yarl==1.7.2
